Create speakers page
A couple of things: 1) I wasn't sure if the order from Eventory was meaningful. I can group people by company or do alphabetical order or whatever if that makes more sense. 2) I googled Ayo Akinyele and it looks like he usually has a J. or Joseph in front of the rest of his name? Not sure if the most prominent person with this name just happens to also be in infosec or something like that.